finished login and register
