+++ frontend/src/pages/Register/Register.jsx
@@ -1,25 +1,24 @@
 import React, { Component } from "react";
-import { Link } from "react-router-dom";
+import { Link, withRouter } from "react-router-dom";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { loginUser } from "../../actions/authActions";
-import { Button, Input , LoginContainer, Wrapper, Label} from './Login.styled'
+import { registerUser } from "../../actions/authActions";
+import { Button, Input , LoginContainer, Wrapper, Label} from './Register.styled'
+import Login from "../Login/Login";
 
-class Login extends Component {
+class Register extends Component {
     constructor() {
         super();
         this.state = {
+            name: "",
             email: "",
             password: "",
+            password2: "",
             errors: {},
         };
     }
 
     componentWillReceiveProps(nextProps) {
-        if (nextProps.auth.isAuthenticated) {
-            this.props.history.push("/home");
-        }
-
         if (nextProps.errors) {
             this.setState({
                 errors: nextProps.errors,
@@ -34,12 +33,14 @@ class Login extends Component {
     onSubmit = (e) => {
         e.preventDefault();
 
-        const userData = {
+        const newUser = {
+            name: this.state.name,
             email: this.state.email,
             password: this.state.password,
+            password2: this.state.password2,
         };
 
-        this.props.loginUser(userData);
+        this.props.registerUser(newUser, this.props.history);
     };
 
     render() {
@@ -47,17 +48,24 @@ class Login extends Component {
 
         return (
             <LoginContainer>
-         
+
+            
                 <div>
                     <Wrapper>
-                        <Label>Sign In</Label>
-
+                        <Label>Register</Label>
                         <form noValidate onSubmit={this.onSubmit}>
                             <div>
-                                <span className="red-text">
-                                    {errors.email}
-                                    {errors.emailnotfound}
-                                </span>
+                                <Input placeholder="Full Name"
+                                    onChange={this.onChange}
+                                    value={this.state.name}
+                                    error={errors.name}
+                                    id="name"
+                                    type="text"
+                                />
+                                
+                                <span className="red-text">{errors.name}</span>
+                            </div>
+                            <div>
                                 <Input placeholder="Email"
                                     onChange={this.onChange}
                                     value={this.state.email}
@@ -65,12 +73,9 @@ class Login extends Component {
                                     id="email"
                                     type="email"
                                 />
+                                <span className="red-text">{errors.email}</span>
                             </div>
                             <div>
-                                <span className="red-text">
-                                    {errors.password}
-                                    {errors.passwordincorrect}
-                                </span>
                                 <Input placeholder="Password"
                                     onChange={this.onChange}
                                     value={this.state.password}
@@ -78,9 +83,20 @@ class Login extends Component {
                                     id="password"
                                     type="password"
                                 />
+                                <span className="red-text">{errors.password}</span>
                             </div>
                             <div>
-                                <Button type="submit">Login</Button>
+                                <Input placeholder="Confirm Password"
+                                    onChange={this.onChange}
+                                    value={this.state.password2}
+                                    error={errors.password2}
+                                    id="password2"
+                                    type="password"
+                                />
+                                <span className="red-text">{errors.password2}</span>
+                            </div>
+                            <div>
+                                <Button type="submit">Sign up</Button>
                             </div>
                         </form>
                     </Wrapper>
@@ -90,8 +106,8 @@ class Login extends Component {
     }
 }
 
-Login.propTypes = {
-    loginUser: PropTypes.func.isRequired,
+Register.propTypes = {
+    registerUser: PropTypes.func.isRequired,
     auth: PropTypes.object.isRequired,
     errors: PropTypes.object.isRequired,
 };
@@ -101,4 +117,4 @@ const mapStateToProps = (state) => ({
     errors: state.errors,
 });
 
-export default connect(mapStateToProps, { loginUser })(Login);
+export default connect(mapStateToProps, { registerUser })(withRouter(Register));
